LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/constraints/form-validation-reportValidity-expected.txt


FAIL [INPUT in TEXT status] no constraint assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in TEXT status] no constraint (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in TEXT status] suffering from being too long assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in TEXT status] suffering from being too long (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in TEXT status] suffering from a pattern mismatch assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in TEXT status] suffering from a pattern mismatch (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in TEXT status] suffering from being missing assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in TEXT status] suffering from being missing (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in SEARCH status] no constraint assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in SEARCH status] no constraint (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in SEARCH status] suffering from being too long assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in SEARCH status] suffering from being too long (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in SEARCH status] suffering from a pattern mismatch assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in SEARCH status] suffering from a pattern mismatch (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in SEARCH status] suffering from being missing assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in SEARCH status] suffering from being missing (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in TEL status] no constraint assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in TEL status] no constraint (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in TEL status] suffering from being too long assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in TEL status] suffering from being too long (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in TEL status] suffering from a pattern mismatch assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in TEL status] suffering from a pattern mismatch (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in TEL status] suffering from being missing assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in TEL status] suffering from being missing (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in PASSWORD status] no constraint assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in PASSWORD status] no constraint (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in PASSWORD status] suffering from being too long assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in PASSWORD status] suffering from being too long (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in PASSWORD status] suffering from a pattern mismatch assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in PASSWORD status] suffering from a pattern mismatch (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in PASSWORD status] suffering from being missing assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in PASSWORD status] suffering from being missing (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in URL status] no constraint assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in URL status] no constraint (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in URL status] suffering from being too long assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in URL status] suffering from being too long (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in URL status] suffering from a pattern mismatch assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in URL status] suffering from a pattern mismatch (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in URL status] suffering from a type mismatch assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in URL status] suffering from a type mismatch (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in URL status] suffering from being missing assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in URL status] suffering from being missing (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in EMAIL status] no constraint assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in EMAIL status] no constraint (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in EMAIL status] suffering from being too long assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in EMAIL status] suffering from being too long (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in EMAIL status] suffering from a pattern mismatch assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in EMAIL status] suffering from a pattern mismatch (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in EMAIL status] suffering from a type mismatch assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in EMAIL status] suffering from a type mismatch (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in EMAIL status] suffering from being missing assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in EMAIL status] suffering from being missing (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in DATETIME status] The datetime type must be supported. assert_equals: The datetime type should be supported. expected "datetime" but got "text"
FAIL [INPUT in DATE status] The date type must be supported. assert_equals: The date type should be supported. expected "date" but got "text"
FAIL [INPUT in MONTH status] The month type must be supported. assert_equals: The month type should be supported. expected "month" but got "text"
FAIL [INPUT in WEEK status] The week type must be supported. assert_equals: The week type should be supported. expected "week" but got "text"
FAIL [INPUT in TIME status] The time type must be supported. assert_equals: The time type should be supported. expected "time" but got "text"
FAIL [INPUT in NUMBER status] suffering from an overflow assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in NUMBER status] suffering from an overflow (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in NUMBER status] suffering from an underflow assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in NUMBER status] suffering from an underflow (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in NUMBER status] suffering from a step mismatch assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in NUMBER status] suffering from a step mismatch (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in NUMBER status] suffering from being missing assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in NUMBER status] suffering from being missing (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in CHECKBOX status] no constraint assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in CHECKBOX status] no constraint (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in CHECKBOX status] suffering from being missing assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in CHECKBOX status] suffering from being missing (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in RADIO status] no constraint assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in RADIO status] no constraint (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in RADIO status] suffering from being missing assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in RADIO status] suffering from being missing (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in FILE status] no constraint assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in FILE status] no constraint (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in FILE status] suffering from being missing assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [INPUT in FILE status] suffering from being missing (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [select]  no constraint assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [select]  no constraint (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [select]  suffering from being missing assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [select]  suffering from being missing (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [textarea]  no constraint assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [textarea]  no constraint (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [textarea]  suffering from being missing assert_true: The reportValidity method doesn't exist. expected true got false
FAIL [textarea]  suffering from being missing (in a form) assert_true: The reportValidity method doesn't exist. expected true got false
